Move ConcertRepository into example.adapter.query package
