+++ sample-app/src/main/scala/example/presentation/MyBoxOfficeResource.scala
@@ -5,8 +5,8 @@ import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.{ ExceptionHandler, Route }
-import example.adapter.ConcertRepository
 import example.adapter.command.BoxOfficeService
+import example.adapter.query.ConcertRepository
 import example.presentation.protocol._
 
 import scala.concurrent._
@@ -21,7 +21,7 @@ final class MyBoxOfficeResource(
 ) extends BoxOfficeResource {
   import SprayJsonSupport._
   import example.adapter.command.BoxOfficeService._
-  import example.adapter.ConcertRepository._
+  import example.adapter.query.ConcertRepository._
   import example.presentation.dsl.ConcertPathMatchers._
 
   // ExceptionHandler を定義する
