+++ sample-app/src/main/scala/example/presentation/protocol/GetConcertsResponseBody.scala
@@ -4,7 +4,7 @@ import spray.json.DefaultJsonProtocol._
 import spray.json.RootJsonFormat
 import ConcertJsonProtocol._
 import example.adapter.ConcertId
-import example.adapter.ConcertRepository.ConcertItem
+import example.adapter.query.ConcertRepository.ConcertItem
 
 /** コンサート一覧取得 レスポンス
   *
