+++ sample-app/src/test/scala/example/presentation/MyBoxOfficeResourceBindSpec.scala
@@ -2,8 +2,8 @@ package example.presentation
 
 import akka.actor.typed.ActorSystem
 import example.ActorSpecBase
-import example.adapter.ConcertRepository
 import example.adapter.command.BoxOfficeService
+import example.adapter.query.ConcertRepository
 import example.testing.tags.ExerciseTest
 import org.mockito.MockitoSugar
 import testkit.AirframeDiSessionSupport
